refactor: common/util/constant -> common/util/enumerate 패키지 이름 변경
