+++ src/main/java/com/airjnc/common/exception/DefaultException.java
@@ -1,6 +1,6 @@
 package com.airjnc.common.exception;
 
-import com.airjnc.common.util.constant.ErrorCode;
+import com.airjnc.common.util.enumerate.ErrorCode;
 import com.airjnc.common.util.factory.ErrorsFactory;
 import org.springframework.validation.Errors;
