+++ src/main/java/com/airjnc/common/exception/NotFoundException.java
@@ -1,21 +1,16 @@
 package com.airjnc.common.exception;
 
 
-import com.airjnc.common.util.constant.ErrorCode;
+import com.airjnc.common.util.enumerate.ErrorCode;
 import com.airjnc.common.util.factory.ErrorsFactory;
-import org.springframework.validation.Errors;
 
 public class NotFoundException extends DefaultException {
 
   public NotFoundException() {
-    super(ErrorsFactory.create(ErrorCode.NOT_FOUND));
+    super(ErrorsFactory.createAndReject(ErrorCode.NOT_FOUND.name()));
   }
 
-  public NotFoundException(Errors errors) {
-    super(errors);
-  }
-
-  public NotFoundException(Errors errors, Throwable cause) {
-    super(errors, cause);
+  public NotFoundException(String objectName, Object[] errorArgs) {
+    super(ErrorsFactory.createAndReject(ErrorCode.NOT_FOUND.name(), objectName, errorArgs));
   }
 }
