+++ src/main/java/com/airjnc/common/exception/UnauthorizedException.java
@@ -1,21 +1,12 @@
 package com.airjnc.common.exception;
 
 
-import com.airjnc.common.util.constant.ErrorCode;
+import com.airjnc.common.util.enumerate.ErrorCode;
 import com.airjnc.common.util.factory.ErrorsFactory;
-import org.springframework.validation.Errors;
 
 public class UnauthorizedException extends DefaultException {
 
   public UnauthorizedException() {
-    super(ErrorsFactory.create(ErrorCode.UNAUTHORIZED));
-  }
-
-  public UnauthorizedException(Errors errors) {
-    super(errors);
-  }
-
-  public UnauthorizedException(Errors errors, Throwable cause) {
-    super(errors, cause);
+    super(ErrorsFactory.createAndReject(ErrorCode.UNAUTHORIZED.name()));
   }
 }
