--- src/main/java/com/airjnc/common/util/constant/SessionKey.java
@@ -1,5 +1,0 @@
-package com.airjnc.common.util.constant;
-
-public enum SessionKey {
-  USER,
-}
