+++ src/main/java/com/airjnc/common/util/enumerate/SessionKey.java
@@ -1,0 +1,5 @@
+package com.airjnc.common.util.enumerate;
+
+public enum SessionKey {
+  USER,
+}
